Specify remaining Perl package versions
These are the version numbers that were involved in the initial Pirana
image build. Aside from better reproducibility, pinning them was also
needed because one of the (latest versions of the) unversioned packages
was giving build errors when rebuilding about one year later.

For "Net::EmptyPort" I did not manage to specify a version number, which
is presumably related to https://github.com/tokuhirom/Test-TCP/issues/63